allow create nested file
